refactor: extract properties logic
